danao data

+++ galaxy-sop-server/src/main/java/com/galaxyinternet/project_danao/service/InfoFromDanaoServiceImpl.java
@@ -629,11 +629,11 @@ public class InfoFromDanaoServiceImpl implements InfoFromDanaoService {
         Long ctimeL = Long.parseLong(ctime);
 
         Long value = curr - ctimeL;
-        if(value.intValue() < 1*60*60*1000){
-        	int ts = value.intValue()/(60*1000);
+        if(value < 1*60*60*1000){
+        	int ts = (int) (value/(60*1000));
 			ctime = ts + "分钟前";
-        }else if(value.intValue() <= 24*60*60*1000 ){
-			int ts = value.intValue()/(60*60*1000);
+        }else if(value <= 24*60*60*1000 ){
+			int ts = (int) (value/(60*60*1000));
 			ctime = ts + "小时前";
 		}else{
         	ctime = DateUtil.longToString(ctimeL);
